requirements.txt
autopep8==2.1.0
absolufy-imports==0.3.1
annotated-types==0.6.0
anyio==4.3.0
aiomysql==0.2.0
aioodbc==0.5.0
aiomqtt==2.3.0
aiosmtplib==3.0.2
aiosqlite
async-timeout==4.0.3
asyncpg==0.29.0
cachetools==5.3.3
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
cx-Oracle==8.3.0
device-detector==5.0.1
dnspython==2.7.0 
ecdsa==0.19.0
email-validator==2.2.0
fastapi==0.110.1
firebase_admin==6.5.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflect==7.2.1
Jinja2==3.1.4
paho-mqtt==2.1.0
pandas==2.2.2
paramiko==3.4.0
plotly==5.22.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.7.0
pydantic_core==2.18.1
pymysql==1.1.1
pyodbc==5.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytest==8.2.0
pytest-dotenv==0.5.2
pytest-tornasync==0.6.0
pytest-asyncio==0.23.6
PyYAML==6.0.1
regex==2023.12.25
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
secure==0.3.0
typing_extensions==4.11.0
uvicorn==0.29.0
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-oauthlib==1.2.0
requests-oauthlib==2.0.0

